+++ src/ganttChartSettingsModels.ts
@@ -218,10 +218,10 @@ export class DaysOffCardSettings extends Card {
         value: dayOfWeekOptions[0]
     });
 
+    topLevelSlice = this.show;
     name: string = "daysOff";
     displayNameKey: string = "Visual_DaysOff";
     slices = [this.fill, this.firstDayOfWeek];
-    topLevelSlice?: formattingSettings.SimpleSlice<any> = this.show;
 }
 
 export class LegendCardSettings extends Card {
@@ -283,6 +283,17 @@ export class LegendCardSettings extends Card {
 }
 
 export class MilestonesCardSettings extends Card {
+    displayDottedLines = new formattingSettings.ToggleSwitch({
+        name: "displayDottedLines",
+        displayNameKey: "Visual_DisplayDottedLines",
+        value: true
+    });
+
+    dottedLinesColor = new formattingSettings.ColorPicker({
+        name: "dottedLinesColor",
+        displayNameKey: "Visual_DottedLinesColor",
+        value: { value: "#cccccc" }
+    });
 
     fill = new formattingSettings.ColorPicker({
         name: "fill",
@@ -299,11 +310,10 @@ export class MilestonesCardSettings extends Card {
 
     name: string = "milestones";
     displayNameKey: string = "Visual_Milestones";
-    slices = [];
+    slices = [this.displayDottedLines, this.dottedLinesColor];
 }
 
 export class TaskLabelsCardSettings extends Card {
-
     show = new formattingSettings.ToggleSwitch({
         name: "show",
         displayNameKey: "Visual_Show",
@@ -340,10 +350,44 @@ export class TaskLabelsCardSettings extends Card {
         }
     });
 
+    collapseAllColor = new formattingSettings.ColorPicker({
+        name: "collapseAllColor",
+        displayNameKey: "Visual_Collapse_All_Color",
+        value: { value: "#000" },
+    });
+
+    collapseAll = new formattingSettings.FontControl({
+        name: "collapseAll",
+        displayNameKey: "Visual_Collapse_All",
+        fontSize: new formattingSettings.NumUpDown({
+            name: "collapseAllFontSize",
+            value: 12,
+            options: {
+                minValue: { value: 1, type: powerbiVisualsApi.visuals.ValidatorType.Min },
+            },
+        }),
+        fontFamily: new formattingSettings.FontPicker({
+            name: "collapseAllFontFamily",
+            value: "Arial, sans-serif"
+        }),
+        bold: new formattingSettings.ToggleSwitch({
+            name: "collapseAllBold",
+            value: false,
+        }),
+        italic: new formattingSettings.ToggleSwitch({
+            name: "collapseAllItalic",
+            value: false,
+        }),
+        underline: new formattingSettings.ToggleSwitch({
+            name: "collapseAllUnderline",
+            value: false,
+        }),
+    });
+
+    topLevelSlice = this.show;
     name: string = "taskLabels";
     displayNameKey: string = "Visual_CategoryLabels";
-    slices = [this.fill, this.fontSize, this.width];
-    topLevelSlice?: formattingSettings.SimpleSlice<any> = this.show;
+    slices = [this.fill, this.fontSize, this.width, this.collapseAllColor, this.collapseAll];
 }
 
 export class TaskCompletionCardSettings extends Card {
@@ -360,10 +404,10 @@ export class TaskCompletionCardSettings extends Card {
         value: undefined
     });
 
+    topLevelSlice= this.show;
     name: string = "taskCompletion";
     displayNameKey: string = "Visual_TaskCompletion";
     slices = [this.maxCompletion];
-    topLevelSlice?: formattingSettings.SimpleSlice<any> = this.show;
 }
 
 export class TooltipConfigCardSettings extends Card {
@@ -452,10 +496,10 @@ export class TaskResourceCardSettings extends Card {
         value: false
     });
 
+    topLevelSlice = this.show;
     name: string = "taskResource";
     displayNameKey: string = "Visual_DataLabels";
     slices = [this.fill, this.fontSize, this.position, this.fullText, this.widthByTask];
-    topLevelSlice?: formattingSettings.SimpleSlice<any> = this.show;
 }
 
 export class DateTypeCardSettings extends Card {
@@ -521,8 +565,8 @@ export class GanttChartSettingsModel extends Model {
     populateMilestones(milestonesWithoutDuplicates: {
         [name: string]: MilestoneDataPoint
     }) {
-        const newSlices = [];
 
+        const newSlices: FormattingSettingsSlice[] = this.milestonesCardSettings.slices;
         if (milestonesWithoutDuplicates) {
             for (const uniqMilestones in milestonesWithoutDuplicates) {
                 const milestone = milestonesWithoutDuplicates[uniqMilestones];